# Conflicts:
#	core/src/main/scala/com/raphtory/api/analysis/graphview/GraphViewImplementation.scala
#	core/src/main/scala/com/raphtory/api/input/Property.scala
#	core/src/main/scala/com/raphtory/internals/components/partition/QueryExecutor.scala
#	core/src/main/scala/com/raphtory/internals/components/querymanager/QueryHandler.scala
#	core/src/main/scala/com/raphtory/internals/management/PythonInterop.scala
#	core/src/main/scala/com/raphtory/internals/management/python/PythonEntrypoint.scala
#	core/src/main/scala/com/raphtory/internals/management/python/UnsafeEmbeddedPythonProxy.scala
#	core/src/main/scala/com/raphtory/internals/management/python/UnsafeGraphBuilder.scala
#	core/src/main/scala/com/raphtory/python/PyRaphtory.scala
#	core/src/test/scala/com/raphtory/facebooktest/FacebookTest.scala
#	examples/enron/src/main/scala/com/raphtory/examples/enron/Runner.scala
#	examples/lotr/src/main/scala/com/raphtory/examples/lotr/LOTRClient.scala
#	examples/lotr/src/main/scala/com/raphtory/examples/lotr/TutorialRunner.scala
#	python/pyraphtory/poetry.lock
#	python/pyraphtory/pyproject.toml
#	python/pyraphtory/pyraphtory/algorithms/sample.py
#	python/pyraphtory/pyraphtory/algorithms/trianglecount.py
#	python/pyraphtory/pyraphtory/builder.py
#	python/pyraphtory/pyraphtory/context.py
#	python/pyraphtory/pyraphtory/graph.py
#	python/pyraphtory/pyraphtory/interop.py
#	python/pyraphtory/pyraphtory/scala/collection.py
#	python/pyraphtory/pyraphtory/vertex.py

+++ python/pyraphtory/pyraphtory/algorithms/trianglecount.py
@@ -1,12 +1,7 @@
-import traceback
-
-from typing import Any
-
 from pyraphtory.algorithm import PyAlgorithm
-from pyraphtory.context import BaseContext
 from pyraphtory.graph import TemporalGraph, Row
-from pyraphtory.vertex import GraphState, Vertex
-from pyraphtory.scala.numeric import Int
+from pyraphtory.vertex import Vertex
+from pyraphtory.scala.implicits.numeric import Long
 
 
 class LocalTriangleCount(PyAlgorithm):
@@ -33,7 +28,7 @@ class LocalTriangleCount(PyAlgorithm):
 class GlobalTriangleCount(LocalTriangleCount):
     def __call__(self, graph: TemporalGraph) -> TemporalGraph:
         return (super().__call__(graph)
-                .set_global_state(lambda s: s.new_adder[Int](name="triangles", retain_state=True))
+                .set_global_state(lambda s: s.new_adder[Long](name="triangles", retain_state=True))
                 .step(lambda v, s: s["triangles"].add(v["triangleCount"])))
 
     def tabularise(self, graph: TemporalGraph):
